+++ lesson6/main.rb
@@ -1,16 +1,25 @@
-require_relative 'manufacturer.rb'
-require_relative 'object_validation.rb'
-require_relative 'instance_counter.rb'
-require_relative 'train.rb'
-require_relative 'train_passenger.rb'
-require_relative 'train_cargo.rb'
-require_relative 'van.rb'
-require_relative 'vans_passenger.rb'
-require_relative 'vans_cargo.rb'
-require_relative 'station.rb'
-require_relative 'route.rb'
+require_relative 'modules/manufacturer'
+require_relative 'modules/object_validation'
+require_relative 'modules/instance_counter'
+require_relative 'station/station'
+require_relative 'station/station_methods'
+require_relative 'route/route'
+require_relative 'route/route_methods'
+require_relative 'train/train_methods'
+require_relative 'train/train'
+require_relative 'train/train_passenger'
+require_relative 'train/train_cargo'
+require_relative 'van/van_methods'
+require_relative 'van/van'
+require_relative 'van/vans_passenger'
+require_relative 'van/vans_cargo'
 
 class Game
+  include StationMethods
+  include RouteMethods
+  include TrainMethods
+  include VanMethods
+
   def initialize
     @stations = []
     @routes = []
@@ -115,370 +124,8 @@ class Game
     pre_path_assignment
   end
 
-  def create_station(name)
-    if @stations.map(&:name).include?(name)
-      'Станция уже существует. Отмена'
-    else
-      @stations << Station.new(name)
-      "Станция #{name} успешно создана!"
-    end
-  end
-
-  def available_stations
-    puts 'Доступные станции: '
-    if @stations.empty?
-      'Список станций пуст. Для начала создайте станцию.'
-    else
-      @stations.map(&:name)
-    end
-  end
-
-  def trains_list_on_station
-    print 'Выберите станцию, для которой хотите просмотреть список поездов: '
-    name = gets.strip.chomp
-    index = stations_index(name)
-    if index.nil?
-      didnt_understand_you
-    else
-      station = @stations[index]
-      puts "Поезда на станции #{station.name}:"
-      station.trains.each { |train| "Поезд #{train.id}, тип: #{train.type}, вагоны: #{train.vans.count} " }
-    end
-  end
-
-  def pre_create_route_actions
-    print 'Введите начальную точку маршрута: '
-    start = gets.strip.chomp
-    print 'Введите конечную точку маршрута: '
-    finish = gets.strip.chomp
-    create_route(start, finish)
-  end
-
-  def create_route(start, finish)
-    stations = @stations.map(&:name)
-    if stations.include?(start) && stations.include?(finish)
-      @routes << Route.new(@routes.empty? ? 1 : @routes[-1].id + 1, start, finish)
-      @routes[-1].waypoints
-    elsif !stations.include?(start)
-      'Начальная точка маршрута не существует!'
-    elsif !stations.include?(finish)
-      'Конечная точка маршрута не существует!'
-    end
-  end
-
-  def route_and_waypoints
-    index = @routes.size
-    index.times do |route_index|
-      print "Маршрут #{route_index + 1} --> "
-      puts @routes[route_index].waypoints.to_s
-    end
-  end
-
-  def pre_edit_route
-    puts 'Доступные маршруты: '
-    route_and_waypoints
-    print 'Введите номер маршрута: '
-    id = gets.to_i
-    if @routes.map(&:id).include?(id)
-      edit_route(id)
-    else
-      didnt_understand_you
-    end
-  end
-
-  def edit_route(id)
-    print 'Вы хотите (д)обавить станцию или (у)далить?: '
-    answer = gets.strip.chomp.downcase
-    add_stations_to_route(id)      if %w[д l].include?(answer)
-    delete_stations_from_route(id) if %w[у e].include?(answer)
-    didnt_understand_you
-  end
-
-  def no_such_station_try_again(id, station)
-    puts "Станция #{station} недоступна для добавления. Попробуете еще раз? (д/н): "
-    answer = gets.strip.chomp.downcase
-    %w[д l].include?(answer) ? add_stations_to_route(id) : 'Ввод отменен пользователем'
-  end
-
-  def add_stations_to_route(id)
-    index = routes_index(id)
-    route = @routes[index]
-    available_stations = @stations.map(&:name) - route.waypoints
-    if available_stations.empty?
-      'Доступных для добавления в этот маршрут станций нет. Выберите другой маршрут или создайте станции.'
-    else
-      print 'Доступные станции: '
-      available_stations.each { |station| print "\'#{station}\' " }
-      puts
-      print 'Введите название станции, которую хотите добавить в маршрут: '
-      station = gets.strip.chomp
-      if available_stations.include?(station)
-        route.add(station)
-        "Станция \'#{station}\' успешно добавлена в маршрут."
-      else
-        no_such_station_try_again(id, station)
-      end
-    end
-  end
-
-  def delete_stations_from_route(id)
-    index = routes_index(id)
-    route = @routes[index]
-    print 'В маршруте есть следующие станции: '
-    route.waypoints.each { |waypoint| print "#{waypoint} " }
-    puts
-    print 'Выберите станцию для удаления: '
-    station = gets.strip.chomp
-    if route.waypoints.include?(station)
-      route.delete(station)
-      "Станция \'#{station}\' успешно удалена из маршрута."
-    else
-      no_such_station_try_again
-    end
-  end
-
-  def pre_path_assignment
-    puts 'Доступные поезда: '
-    @trains.each do |train|
-      train.route ? (puts "Поезд #{train.id}->#{train.route} ") : (puts "Поезд #{train.id}-> маршрут не назначен ")
-    end
-    puts 'Доступные маршруты: '
-    @routes.each { |route| puts "Маршрут #{route.id}->#{route.waypoints} " }
-
-    puts 'Введите номер поезда и номер маршрута для него: '
-    print 'Номер поезда: '
-    train_id = gets.strip.chomp
-    print 'Номер маршрута: '
-    route_id = gets.to_i
-
-    if trains_include?(train_id) && @routes.map(&:id).include?(route_id)
-      path_assignment(train_id, route_id)
-    else
-      'Вы неверно ввели номер поезда или номер маршрута.'
-    end
-  end
-
-  def path_assignment(train_id, route_id)
-    index = trains_map_index(train_id)
-    train = @trains[index]
-
-    index = routes_index(route_id)
-    route = @routes[index]
-    train.route = route.waypoints
-
-    index = stations_index(train.route.first)
-    station = @stations[index]
-    message = station.train_arrival(train)
-
-    "Маршрут ##{route_id} успешно добавлен к поезду #{train_id}. " + message
-  end
-
-  def create_train_exceptions_handling
-    begin
-      puts first_step_to_create_train
-    rescue ArgumentError => e
-      puts_with_effects(e.to_s)
-      puts 'Попробуем еще раз...'
-      retry
-    end
-  end
-
-  def first_step_to_create_train
-    print 'Какой поезд хотите создать - (п)ассажирский или (г)рузовой?: '
-    type = gets.strip.chomp.downcase
-    second_step_to_create_train(type)
-  end
-
-  def second_step_to_create_train(type)
-    puts 'Придумайте номер / идентификатор для поезда.'
-    puts 'Формат - три буквы или цифры в любом порядке, дефис - по желанию,'
-    print 'и еще 2 буквы или цифры после дефиса: '
-    id = gets.strip.chomp
-    if trains_include?(id)
-      "Поезд #{id} уже существует. Придумайте другой. Отмена."
-    else
-      print 'Укажите производителя поезда: '
-      manufacturer = gets.strip.chomp
-      create_train(id, type, manufacturer)
-    end
-  end
-
-  def create_train(id, type, manufacturer)
-    if %w[п g].include?(type)
-      @trains << PassengerTrain.new(id, manufacturer)
-      "Пассажирский поезд #{id} создан!"
-    elsif %w[г u].include?(type)
-      @trains << CargoTrain.new(id, manufacturer)
-      "Грузовой поезд #{id} создан!"
-    else
-    raise ArgumentError, 'Проверьте правильность ввода типа поезда'
-    end
-  end
-
-  def where_to_move_train(id)
-    print 'Куда двигать поезд - (в)перед или (н)азад: '
-    choise = gets.strip.chomp.downcase
-    if %w[в d н y].include?(choise)
-      move(id, choise)
-    else
-      didnt_understand_you
-    end
-  end
-
-  def move(id, choise)
-    index = trains_map_index(id)
-    train = @trains[index]
-    departure = train.now_station
-    arrival = train.move_forward if %w[в d].include?(choise)
-    arrival = train.move_backward if %w[н y].include?(choise)
-    departure_arrival(train, departure, arrival)
-  end
-
-  def departure_arrival(train, departure, arrival)
-    index = stations_index(arrival)
-    if index.nil?
-      arrival
-    else
-      index = stations_index(departure)
-      station = @stations[index]
-      station.train_departure(train)
-
-      index = stations_index(arrival)
-      station = @stations[index]
-      station.train_arrival(train)
-    end
-  end
-
-  def create_van
-    puts 'Какой вагон Вы хотите создать?'
-    print '(п)ассажирский или (г)рузовой: '
-    type = gets.strip.chomp.downcase
-    print 'Вид - Спальный (СВ), Купейный (КВ): ' if %w[п g].include?(type)
-    print 'Вид - Угольный, Зерновой, Цистерна: ' if %w[г u].include?(type)
-    kind = gets.strip.chomp
-    print 'Укажите производитя вагона: '
-    manufacturer = gets.strip.chomp
-    adding_to_vans(type, kind, manufacturer)
-  end
-
-  def adding_to_vans(type, kind, manufacturer)
-    number = @vans.empty? ? 1 : @vans[-1].number + 1
-    if %w[п g].include?(type)
-      @vans << PassengerVan.new(number, kind, manufacturer)
-      "Пассажиркий вагон ##{@vans[-1].number} создан."
-    elsif %w[г u].include?(type)
-      @vans << CargoVan.new(number, kind, manufacturer)
-      "Грузовой вагон ##{@vans[-1].number} создан"
-    else
-      'Такого типа вагонов не существует. Попробуйте еще раз. Отмена.'
-    end
-  end
-
-  def pre_add_van
-    puts 'Прицепляем новые вагоны...'
-    available_trains
-    puts 'Доступные вагоны:'
-    @vans.each { |van| puts "##{van.number}. Тип: #{van.type}, вид: #{van.kind}. " if van.status == 'free' }
-    print 'Введите номер поезда: '
-    id = gets.to_i
-    print 'Введите номер вагона: '
-    number = gets.to_i
-    if trains_include?(id) && @vans.map(&:number).include?(number)
-      add_van(id, number)
-    else
-      didnt_understand_you
-    end
-  end
-
-  def add_van(id, number)
-    index = @vans.map(&:number).index number
-    van = @vans[index]
-    index = trains_map_index(id)
-    train = @trains[index]
-    train.add_van(van)
-  end
-
-  def pre_delete_van
-    available_trains
-    print 'Введите номер поезда, у которого Вы хотите отцепить вагон: '
-    id = gets.to_i
-    if trains_include?(id)
-      index = trains_map_index(id)
-      train = @trains[index]
-      delete_van(train)
-    else
-      no_such_train
-    end
-  end
-
-  def delete_van(train)
-    if train.vans.count.zero?
-      'Вагонов нет. Удалять нечего.'
-    else
-      puts "Поезд #{train.id}. Тип: #{train.type}, скорость: #{train.speed}, имеет следующие вагоны:"
-      train.vans.each { |van| puts "< Вагон ##{van.number}, \'#{van.type}\', \'#{van.kind}\' >" }
-      van_choise(train)
-    end
-  end
-
-  def van_choise(train)
-    print 'Выберите номер вагона для отцепления: '
-    number = gets.to_i
-    if @vans.map(&:number).include?(number)
-      index = @vans.map(&:number).index number
-      van = @vans[index]
-      train.delete_van(van)
-    else
-      didnt_understand_you
-    end
-  end
-
-  def change_speed(id, choise)
-    index = trains_map_index(id)
-    train = @trains[index]
-    print 'Введите скорость: '
-    speed = gets.to_i
-    train.speed_up(speed) if choise == 'up'
-    train.speed_down(speed) if choise == 'down'
-  end
-
-  def stations_index(name)
-    @stations.map(&:name).index name
-  end
-
-  def train_stop(id)
-    @trains.each { |train| train.stop if train.id == id }
-  end
-
-  def trains_include?(id)
-    @trains.map(&:id).include?(id)
-  end
-
-  def trains_map_index(id)
-    @trains.map(&:id).index id
-  end
-
-  def available_trains
-    puts 'Доступные поезда:'
-    @trains.each { |train| puts "Поезд #{train.id}. Тип: #{train.type}, вагонов: #{train.vans.count}, скорость: #{train.speed}."}
-  end
-
-  def train_choise
-    available_trains
-    print 'Введите номер поезда: '
-    @id = gets.strip.chomp
-  end
-
-  def routes_index(id)
-    @routes.map(&:id).index id
-  end
-
   def didnt_understand_you
     'Ввод не распознан. Попробуйте еще раз. Отмена.'
-  end
-
-  def no_such_train(id)
-    "Нет такого поезда - #{id}"
   end
 
   def puts_with_effects(string)
