+++ lesson6/modules/instance_counter.rb
