+++ lesson6/modules/manufacturer.rb
