+++ lesson6/modules/object_validation.rb
