+++ lesson6/station/station.rb
