+++ lesson6/train/train.rb
