+++ lesson6/train/train_cargo.rb
