+++ lesson6/train/train_passenger.rb
