+++ lesson6/van/van.rb
