+++ lesson6/van/vans_cargo.rb
