+++ lesson6/van/vans_passenger.rb
